# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CRM monitoring API for arbitrage trading system.",
        "title": "Arbitrage VM CRM API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/pnl/by-component": {
            "get": {
                "description": "Summarize PnL grouped by funding, trading_fee, and trading_pnl",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PnL"
                ],
                "summary": "PnL by component",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by exchange. Supports comma-separated values.",
                        "name": "exchange",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by base asset. Supports comma-separated values.",
                        "name": "base",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by quote asset. Supports comma-separated values.",
                        "name": "quote",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by component: funding, trading_fee, trading_pnl. Supports comma-separated values.",
                        "name": "component",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source type. Supports comma-separated values.",
                        "name": "source_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source id. Supports comma-separated values.",
                        "name": "source_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive start time. Supports RFC3339 or YYYY-MM-DD.",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exclusive end time. Supports RFC3339 or YYYY-MM-DD.",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum rows to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PnLGroupList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/pnl/by-exchange": {
            "get": {
                "description": "Summarize PnL grouped by exchange",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PnL"
                ],
                "summary": "PnL by exchange",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by exchange. Supports comma-separated values.",
                        "name": "exchange",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by base asset. Supports comma-separated values.",
                        "name": "base",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by quote asset. Supports comma-separated values.",
                        "name": "quote",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by component: funding, trading_fee, trading_pnl. Supports comma-separated values.",
                        "name": "component",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source type. Supports comma-separated values.",
                        "name": "source_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source id. Supports comma-separated values.",
                        "name": "source_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive start time. Supports RFC3339 or YYYY-MM-DD.",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exclusive end time. Supports RFC3339 or YYYY-MM-DD.",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum rows to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PnLGroupList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/pnl/by-pair": {
            "get": {
                "description": "Summarize PnL grouped by base and quote",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PnL"
                ],
                "summary": "PnL by pair",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by exchange. Supports comma-separated values.",
                        "name": "exchange",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by base asset. Supports comma-separated values.",
                        "name": "base",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by quote asset. Supports comma-separated values.",
                        "name": "quote",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by component: funding, trading_fee, trading_pnl. Supports comma-separated values.",
                        "name": "component",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source type. Supports comma-separated values.",
                        "name": "source_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source id. Supports comma-separated values.",
                        "name": "source_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive start time. Supports RFC3339 or YYYY-MM-DD.",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exclusive end time. Supports RFC3339 or YYYY-MM-DD.",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum rows to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PnLGroupList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/pnl/events": {
            "get": {
                "description": "List persisted PnL ledger events",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PnL"
                ],
                "summary": "PnL events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by exchange. Supports comma-separated values.",
                        "name": "exchange",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by base asset. Supports comma-separated values.",
                        "name": "base",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by quote asset. Supports comma-separated values.",
                        "name": "quote",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by component: funding, trading_fee, trading_pnl. Supports comma-separated values.",
                        "name": "component",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source type. Supports comma-separated values.",
                        "name": "source_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source id. Supports comma-separated values.",
                        "name": "source_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive start time. Supports RFC3339 or YYYY-MM-DD.",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exclusive end time. Supports RFC3339 or YYYY-MM-DD.",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum rows to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PnLEventList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/pnl/summary": {
            "get": {
                "description": "Summarize PnL totals and grouped breakdowns",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PnL"
                ],
                "summary": "PnL summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by exchange. Supports comma-separated values.",
                        "name": "exchange",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by base asset. Supports comma-separated values.",
                        "name": "base",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by quote asset. Supports comma-separated values.",
                        "name": "quote",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by component: funding, trading_fee, trading_pnl. Supports comma-separated values.",
                        "name": "component",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source type. Supports comma-separated values.",
                        "name": "source_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source id. Supports comma-separated values.",
                        "name": "source_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive start time. Supports RFC3339 or YYYY-MM-DD.",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exclusive end time. Supports RFC3339 or YYYY-MM-DD.",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum grouped rows per breakdown",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PnLSummary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.PnLEvent": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "base": {
                    "type": "string"
                },
                "component": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "event_time": {
                    "type": "string"
                },
                "exchange": {
                    "type": "string"
                },
                "pair": {
                    "type": "string"
                },
                "quote": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                }
            }
        },
        "response.PnLEventList": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.PnLEvent"
                    }
                },
                "limit": {
                    "type": "integer"
                }
            }
        },
        "response.PnLGroup": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "string"
                },
                "component": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "exchange": {
                    "type": "string"
                },
                "first_event_time": {
                    "type": "string"
                },
                "funding_amount": {
                    "type": "string"
                },
                "last_event_time": {
                    "type": "string"
                },
                "pair": {
                    "type": "string"
                },
                "quote": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "string"
                },
                "trading_fee_amount": {
                    "type": "string"
                },
                "trading_pnl_amount": {
                    "type": "string"
                }
            }
        },
        "response.PnLGroupList": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.PnLGroup"
                    }
                },
                "limit": {
                    "type": "integer"
                }
            }
        },
        "response.PnLSummary": {
            "type": "object",
            "properties": {
                "by_component": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.PnLGroup"
                    }
                },
                "by_exchange": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.PnLGroup"
                    }
                },
                "by_pair": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.PnLGroup"
                    }
                },
                "count": {
                    "type": "integer"
                },
                "funding_amount": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "string"
                },
                "trading_fee_amount": {
                    "type": "string"
                },
                "trading_pnl_amount": {
                    "type": "string"
                }
            }
        },
...
    }
}